+++ global_data.py
@@ -1,5 +1,6 @@
 from wipyWS2812.ws2812 import WS2812
 
+DEBUG = True
 INADDR_ANY = 0
 global_epoch = 0  # time over ntp-server
