+++ main.py
@@ -25,479 +25,12 @@ For a complete discussion, see http://www.makermusings.com
 More info about the MicroPython Version see https://lemariva.com
 
 '''
-import gc
-import machine
-import network
-import time
-
 try:
     import _thread
-
     thread_available = True
 except:
     thread_available = False
-
-try:
-    import uselect as select
-except:
-    import socket
-try:
-    import usocket as socket
-except:
-    import socket
-
-# for ws2812b
-from wipyWS2812.ws2812 import WS2812
-from uos import uname
-
-try:
-    from machine import RTC
-
-    if not hasattr(RTC, "ntp_sync"):
-        from ESP32MicroPython.timeutils import RTC
-except:
-    from ESP32MicroPython.timeutils import RTC
-
-from xmls import SETUP_XML, EVENT_SERVICE_XML, GET_BINARY_STATE_SOAP
-from helpers import dbg, inet_aton, format_timetuple_and_zone
-from global_data import *
-from rest_api_handler import RestApiHandler
-from upnp_broadcast_responder import UpnpBroadcastResponder
-
-class Poller:
-    # A simple utility class to wait for incoming data to be
-    # read on a socket.
-    def __init__(self):
-        if "poll" in dir(select):
-            self.use_poll = True
-            self.poller = select.poll()
-        else:
-            self.use_poll = False
-        self.targets = {}
-
-    def add(self, target, socket=None):
-        if not socket:
-            socket = target.sockets()
-        if self.use_poll:
-            self.poller.register(socket, select.POLLIN)
-        # dbg("add device on fileno: %s" % socket.fileno() )
-        self.targets[socket.fileno()] = target
-        # dbg("size targets: %s" % len(self.targets))
-
-    def remove(self, target, socket=None):
-        if not socket:
-            socket = target.sockets()
-        if self.use_poll:
-            self.poller.unregister(socket)
-        # dbg("remove device on fileno: %s" % socket.fileno() )
-        gc.collect()
-
-    def poll(self, timeout=100):
-        if self.use_poll:
-            ready = self.poller.poll(timeout)
-        else:
-            ready = []
-            if len(self.targets) > 0:
-                (rlist, wlist, xlist) = select.select(
-                    self.targets.keys(), [], [], timeout
-                )
-                ready = [(x, None) for x in rlist]
-
-        for one_ready in ready:
-            target = self.targets.get(one_ready[0].fileno(), None)
-            dbg("Targets %s" % str(self.targets.keys()))
-            if target:
-                # dbg("get socket with fileno: %s" % str(one_ready[0].fileno()) +  " len: %s" % len(one_ready) + " selected: %s " % str(target.fileno()) )
-                # update time
-                target.do_read(one_ready[0])
-
-
-class Upnp_device:
-    '''
-     Base class for a generic UPnP device. This is far from complete
-     but it supports either specified or automatic IP address and port
-     selection.
-    '''
-
-    this_host_ip = None
-
-    @staticmethod
-    def local_ip_address():
-        if not Upnp_device.this_host_ip:
-            try:
-                ap_if = network.WLAN()
-                Upnp_device.this_host_ip = ap_if.ifconfig()[0]
-            except:
-                Upnp_device.this_host_ip = "127.0.0.1"
-            dbg("got local address of %s" % Upnp_device.this_host_ip)
-        return Upnp_device.this_host_ip
-
-    def __init__(
-        self,
-        listener,
-        poller,
-        port,
-        root_url,
-        server_version,
-        persistent_uuid,
-        other_headers=None,
-        ip_address=None,
-    ):
-        self.listener = listener
-        self.poller = poller
-        self.port = port
-        self.root_url = root_url
-        self.server_version = server_version
-        self.persistent_uuid = persistent_uuid
-        self.uuid = machine.unique_id()
-        self.other_headers = other_headers
-
-        if ip_address:
-            self.ip_address = ip_address
-        else:
-            self.ip_address = Upnp_device.local_ip_address()
-
-        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.socket.bind((self.ip_address, self.port))
-        self.socket.listen(5)
-        if self.port == 0:
-            self.port = self.socket.getsockname()[1]
-        self.poller.add(self)
-        self.client_sockets = {}
-        self.listener.add_device(self)
-
-    def fileno(self):
-        return self.socket.fileno()
-
-    def sockets(self):
-        return self.socket
-
-    def do_read(self, socket):
-        fileno = socket.fileno()
-        dbg("Fileno %s" % fileno + " socket fileno %s" % self.socket.fileno())
-
-        if fileno == self.socket.fileno():
-            try:
-                (client_socket, client_address) = self.socket.accept()
-                self.poller.add(self, client_socket)
-                self.client_sockets[client_socket.fileno()] = client_socket
-            except Exception as e:
-                dbg("################################## Socket busy! %s" % str(e))
-        else:
-            data, sender = self.client_sockets[fileno].recvfrom(4096)
-            if not data:
-                self.poller.remove(self, self.client_sockets[fileno])
-                self.client_sockets[fileno].close()
-
-            else:
-                dbg("send response to socket!: %s" % str(fileno))
-                self.handle_request(data, sender, self.client_sockets[fileno])
-            gc.collect()
-
-    def handle_request(self, data, sender, socket):
-        pass
-
-    def get_name(self):
-        return "unknown"
-
-    def respond_to_search(self, destination, search_target):
-        dbg("Responding to search for %s" % self.get_name())
-        date_str = format_timetuple_and_zone(clock.gmtime(), "GMT")
-        location_url = self.root_url % {
-            "ip_address": self.ip_address,
-            "port": self.port,
-        }
-        message = (
-            "HTTP/1.1 200 OK\r\n"
-            "CACHE-CONTROL: max-age=86400\r\n"
-            "DATE: %s\r\n"
-            "EXT:\r\n"
-            "LOCATION: %s\r\n"
-            'OPT: "http://schemas.upnp.org/upnp/1/0/"; ns=01\r\n'
-            "01-NLS: %s\r\n"
-            "SERVER: %s\r\n"
-            "ST: %s\r\n"
-            "USN: uuid:%s::%s\r\n"
-            "X-User-Agent: redsonic\r\n\r\n"
-            % (
-                date_str,
-                location_url,
-                self.uuid,
-                self.server_version,
-                search_target,
-                self.persistent_uuid,
-                search_target,
-            )
-        )
-
-        if self.other_headers:
-            for header in self.other_headers:
-                message += "%s\r\n" % header
-        message += "\r\n"
-
-        try:
-            temp_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            temp_socket.sendto(message, destination)
-            temp_socket.close()
-            gc.collect()
-        except Exception as e:
-            dbg("Got problem to send response %s" % str(e))
-
-
-class Fauxmo(Upnp_device):
-    '''
-     This subclass does the bulk of the work to mimic a WeMo switch on the network.
-    '''
-
-    @staticmethod
-    def make_uuid(name):
-        return "".join(
-            ["%x" % sum([ord(c) for c in name])]
-            + ["%x" % ord(c) for c in "%sFauxmo!" % name]
-        )[:14]
-
-    def __init__(self, name, listener, poller, ip_address, port, action_handler=None):
-        self.serial = self.make_uuid(name)
-        self.name = name
-        self.ip_address = ip_address
-        self.relayState = 0
-        persistent_uuid = "Socket-1_0-" + self.serial
-        other_headers = ["X-User-Agent: redsonic"]
-        Upnp_device.__init__(
-            self,
-            listener,
-            poller,
-            port,
-            "http://%(ip_address)s:%(port)s/setup.xml",
-            "Unspecified, UPnP/1.0, Unspecified",
-            persistent_uuid,
-            other_headers=other_headers,
-            ip_address=ip_address,
-        )
-        if action_handler:
-            self.action_handler = action_handler
-        else:
-            self.action_handler = self
-        dbg(
-            "Fauxmo device '%s' ready on %s:%s"
-            % (self.name, self.ip_address, self.port)
-        )
-
-    def get_name(self):
-        return self.name
-
-    def handle_request(self, data, sender, socket):
-        if (
-            data.find(b"POST /upnp/control/basicevent1 HTTP/1.1") == 0
-            and data.find(b"urn:Belkin:service:basicevent:1#GetBinaryState") != -1
-        ):
-            state = self.getState()
-            soap = GET_BINARY_STATE_SOAP % {"state_realy": state}
-            date_str = format_timetuple_and_zone(clock.gmtime(), "GMT")
-            message = (
-                "HTTP/1.1 200 OK\r\n"
-                "CONTENT-LENGTH: %d\r\n"
-                'CONTENT-TYPE: text/xml charset="utf-8"\r\n'
-                "DATE: %s\r\n"
-                "EXT:\r\n"
-                "SERVER: Unspecified, UPnP/1.0, Unspecified\r\n"
-                "X-User-Agent: redsonic\r\n"
-                "CONNECTION: close\r\n"
-                "\r\n"
-                "%s" % (len(soap), date_str, soap)
-            )
-            socket.send(message)
-        elif data.find(b"GET /eventservice.xml HTTP/1.1") == 0:
-            dbg("Responding to eventservice.xml for %s" % self.name)
-            date_str = format_timetuple_and_zone(clock.gmtime(), "GMT")
-            message = (
-                "HTTP/1.1 200 OK\r\n"
-                "CONTENT-LENGTH: %d\r\n"
-                "CONTENT-TYPE: text/xml\r\n"
-                "DATE: %s\r\n"
-                "LAST-MODIFIED: Sat, 01 Jan 2000 00:01:15 GMT\r\n"
-                "SERVER: Unspecified, UPnP/1.0, Unspecified\r\n"
-                "X-User-Agent: redsonic\r\n"
-                "CONNECTION: close\r\n"
-                "\r\n"
-                "%s" % (len(EVENT_SERVICE_XML), date_str, EVENT_SERVICE_XML)
-            )
-            socket.send(message)
-        elif data.find(b"GET /setup.xml HTTP/1.1") == 0:
-            dbg("Responding to setup.xml for %s" % self.name)
-            xml = SETUP_XML % {"device_name": self.name, "device_serial": self.serial}
-            date_str = format_timetuple_and_zone(clock.gmtime(), "GMT")
-            message = (
-                "HTTP/1.1 200 OK\r\n"
-                "CONTENT-LENGTH: %d\r\n"
-                "CONTENT-TYPE: text/xml\r\n"
-                "DATE: %s\r\n"
-                "LAST-MODIFIED: Sat, 01 Jan 2000 00:01:15 GMT\r\n"
-                "SERVER: Unspecified, UPnP/1.0, Unspecified\r\n"
-                "X-User-Agent: redsonic\r\n"
-                "CONNECTION: close\r\n"
-                "\r\n"
-                "%s" % (len(xml), date_str, xml)
-            )
-            socket.send(message)
-        elif (
-            data.find(b'SOAPACTION: "urn:Belkin:service:basicevent:1#SetBinaryState"')
-            != -1
-        ):
-            success = False
-            if data.find(b"<BinaryState>1</BinaryState>") != -1:
-                # on
-                dbg("Responding to ON for %s" % self.name)
-                self.relayState = 1
-                success = self.action_handler.on()
-            elif data.find(b"<BinaryState>0</BinaryState>") != -1:
-                # off
-                dbg("Responding to OFF for %s" % self.name)
-                self.relayState = 0
-                success = self.action_handler.off()
-            else:
-                dbg("Unknown Binary State request:")
-
-            if success:
-                state = self.getState()
-                soap = GET_BINARY_STATE_SOAP % {"state_realy": state}
-                date_str = format_timetuple_and_zone(clock.gmtime(), "GMT")
-                message = (
-                    "HTTP/1.1 200 OK\r\n"
-                    "CONTENT-LENGTH: %d\r\n"
-                    'CONTENT-TYPE: text/xml charset="utf-8"\r\n'
-                    "DATE: %s\r\n"
-                    "EXT:\r\n"
-                    "SERVER: Unspecified, UPnP/1.0, Unspecified\r\n"
-                    "X-User-Agent: redsonic\r\n"
-                    "CONNECTION: close\r\n"
-                    "\r\n"
-                    "%s" % (len(soap), date_str, soap)
-                )
-                socket.send(message)
-        else:
-            dbg(data)
-
-    def on(self):
-        return False
-
-    def off(self):
-        return True
-
-    def getState(self):
-        return self.relayState
-
-
-
-class InvalidPortException(Exception):
-    # Exception definitions that are used in the package
-    pass
-
-
-def thread_echo(args):
-    global DEBUG
-    global clock
-    global ws2812_chain
-    # Set up our singleton for polling the sockets for data ready
-
-    ws2812_chain = WS2812(ledNumber=ledNumber, brightness=100)
-    p = Poller()
-
-    '''
-     Each entry is a list with the following elements:
-
-     # name of the virtual switch
-     # handler object with 'on' and 'off' methods (e.g. RestApiHandler((rrr, ggg, bbb), lux)})
-     # port #
-
-     NOTE: As of 2015-08-17, the Echo appears to have a hard-coded limit of
-## ftp access
-#from ftp import ftpserver
-
-     16 switches it can control. Only the first 16 elements of the 'devices'
-     list will be used.
-     NOTE: Micropython has a limitation in the number of opened sockets (8).
-     Then, the maximal device number is limited to 3.
-    '''
-    devices = [
-        {
-            "description": "white led",
-            "port": 12340,
-            "handler": RestApiHandler((255, 255, 255), 50),
-        },
-        {
-            "description": "red led",
-            "port": 12341,
-            "handler": RestApiHandler((255, 0, 0), 50),
-        },
-        {
-            "description": "blue led",
-            "port": 12342,
-            "handler": RestApiHandler((30, 144, 255), 90),
-        },
-        {
-            "description": "green led",
-            "port": 12343,
-            "handler": RestApiHandler((0, 255, 0), 90),
-        },
-        {
-            "description": "orange led",
-            "port": 12344,
-            "handler": RestApiHandler((255, 165, 0), 90),
-        },
-    ]
-
-    # Set up our singleton listener for UPnP broadcasts
-    u = UpnpBroadcastResponder()
-    u.init_socket()
-    # Add the UPnP broadcast listener to the poller so we can respond
-    # when a broadcast is received.
-    p.add(u)
-
-    # Create our Fauxmo virtual switch devices
-    # Initialize Fauxmo devices
-    for device in devices:
-        # if `port` doesn't exist, populate it
-        # if it isnt an int, flip out with a descriptive exception
-        if not device.get("port"):
-            device["port"] = 0
-        elif type(device["port"]) is not int:
-            raise InvalidPortException(
-                "Invalid port of type: {}, with a value of: {}".format(
-                    type(device["port"]), device["port"]
-                )
-            )
-        Fauxmo(
-            device["description"],
-            u,
-            p,
-            None,
-            device["port"],
-            action_handler=device["handler"],
-        )
-
-    # setting the clock using ntp
-    if uname().machine == "WiPy with ESP32":
-        # Wipy 2.0
-        clock_tmp = RTC()
-        clock_tmp.ntp_sync("time1.google.com")
-        clock = time  # gmtime function needed
-    elif uname().machine == "ESP32 module with ESP32":
-        # Wemos ESP-WROOM-32
-        clock = RTC()  # gmtime function needed
-        clock.ntp_sync("time1.google.com")
-
-    dbg("Entering main loop\n")
-    while True:
-        try:
-            # Allow time for a ctrl-c to stop the process
-            p.poll(10)
-            time.sleep(0.1)
-            gc.collect()
-        except Exception as e:
-            dbg(e)
-            # break
-
+from app import thread_echo
 
 if thread_available:
     print("Starting echo serviceList on separated thread\n")
